Refactor code
refs #7
